requirements.txt
PyAutoGUI==0.9.54
numpy==2.0.0
pillow==10.4.0
selenium==4.22.0